+++ projects/natural/src/lib/modules/alert/public-api.ts
@@ -2,6 +2,6 @@
  * Public API Surface of natural
  */
 
-export * from './alert.module';
+export * from '../../natural.module';
 export * from './alert.service';
 export * from './confirm.component';
