+++ projects/natural/src/lib/modules/columns-picker/public-api.ts
@@ -4,4 +4,3 @@
 
 export * from './columns-picker-column.directive';
 export * from './columns-picker.component';
-export * from './columns-picker.module';
