+++ projects/natural/src/lib/modules/detail-header/public-api.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './detail-header.component';
-export * from './detail-header.module';
