--- projects/natural/src/lib/modules/dialog-trigger/dialog-trigger.module.ts
@@ -1,11 +1,0 @@
-import {CommonModule} from '@angular/common';
-import {NgModule} from '@angular/core';
-import {MatDialogModule} from '@angular/material/dialog';
-import {NaturalDialogTriggerComponent} from './dialog-trigger.component';
-
-@NgModule({
-    declarations: [NaturalDialogTriggerComponent],
-    imports: [CommonModule, MatDialogModule],
-    exports: [NaturalDialogTriggerComponent],
-})
-export class NaturalDialogTriggerModule {}
