+++ projects/natural/src/lib/modules/dialog-trigger/public-api.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './dialog-trigger.component';
-export * from './dialog-trigger.module';
