+++ projects/natural/src/lib/modules/file/file-select.directive.spec.ts
@@ -1,7 +1,7 @@
 import {ViewChild, Component, NgModule} from '@angular/core';
 import {ComponentFixture, TestBed, waitForAsync} from '@angular/core/testing';
 import {NaturalFileSelectDirective} from './file-select.directive';
-import {NaturalFileModule} from './file.module';
+import {NaturalModule} from '@ecodev/natural';
 
 @Component({
     template: '<input type="file" naturalFileSelect />',
@@ -11,7 +11,7 @@ export class ContainerComponent {
 }
 
 @NgModule({
-    imports: [NaturalFileModule],
+    imports: [NaturalModule],
     declarations: [ContainerComponent],
 })
 export class AppModule {}
