+++ projects/natural/src/lib/modules/file/public-api.ts
@@ -5,6 +5,5 @@
 export * from './file-drop.directive';
 export * from './file-select.directive';
 export * from './file.service';
-export {InvalidFile, FileSelection} from './abstract-file';
 export * from './component/file.component';
-export * from './file.module';
+export {InvalidFile, FileSelection} from './abstract-file';
