--- projects/natural/src/lib/modules/fixed-button-detail/fixed-button-detail.module.ts
@@ -1,12 +1,0 @@
-import {CommonModule} from '@angular/common';
-import {NgModule} from '@angular/core';
-import {NaturalFixedButtonModule} from '../fixed-button/fixed-button.module';
-import {NaturalFixedButtonDetailComponent} from './fixed-button-detail.component';
-import {MatTooltipModule} from '@angular/material/tooltip';
-
-@NgModule({
-    declarations: [NaturalFixedButtonDetailComponent],
-    imports: [CommonModule, NaturalFixedButtonModule, MatTooltipModule],
-    exports: [NaturalFixedButtonDetailComponent],
-})
-export class NaturalFixedButtonDetailModule {}
