+++ projects/natural/src/lib/modules/fixed-button-detail/public-api.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './fixed-button-detail.component';
-export * from './fixed-button-detail.module';
