+++ projects/natural/src/lib/modules/fixed-button/fixed-button.component.spec.ts
@@ -1,8 +1,6 @@
 import {waitForAsync, ComponentFixture, TestBed} from '@angular/core/testing';
 import {RouterTestingModule} from '@angular/router/testing';
-import {NaturalIconModule} from '../icon/icon.module';
 import {NaturalFixedButtonComponent} from './fixed-button.component';
-import {NaturalFixedButtonModule} from './fixed-button.module';
 
 describe('FixedButtonComponent', () => {
     let component: NaturalFixedButtonComponent;
@@ -12,7 +10,7 @@ describe('FixedButtonComponent', () => {
         waitForAsync(() => {
             TestBed.configureTestingModule({
                 declarations: [],
-                imports: [RouterTestingModule, NaturalIconModule.forRoot({}), NaturalFixedButtonModule],
+                imports: [RouterTestingModule],
             }).compileComponents();
         }),
     );
