+++ projects/natural/src/lib/modules/fixed-button/public-api.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './fixed-button.component';
-export * from './fixed-button.module';
