+++ projects/natural/src/lib/modules/icon/public-api.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './icon.component';
-export * from './icon.module';
