--- projects/natural/src/lib/modules/panels/panels.module.ts
@@ -1,25 +1,0 @@
-import {CommonModule} from '@angular/common';
-import {ModuleWithProviders, NgModule} from '@angular/core';
-import {MatDialogModule} from '@angular/material/dialog';
-import {RouterModule} from '@angular/router';
-import {NaturalPanelsComponent} from './panels.component';
-import {NaturalPanelsHooksConfig, PanelsHooksConfig} from './types';
-
-@NgModule({
-    declarations: [NaturalPanelsComponent],
-    imports: [CommonModule, RouterModule, MatDialogModule],
-    exports: [NaturalPanelsComponent],
-})
-export class NaturalPanelsModule {
-    public static forRoot(hooks?: NaturalPanelsHooksConfig): ModuleWithProviders<NaturalPanelsModule> {
-        return {
-            ngModule: NaturalPanelsModule,
-            providers: [
-                {
-                    provide: PanelsHooksConfig,
-                    useValue: hooks,
-                },
-            ],
-        };
-    }
-}
