+++ projects/natural/src/lib/modules/panels/public-api.ts
@@ -2,7 +2,6 @@
  * Public API Surface of natural
  */
 
-export * from './panels.module';
 export * from './panels.component';
 export * from './panels.service';
 export {naturalPanelsUrlMatcher} from './panels.urlmatcher';
