+++ projects/natural/src/lib/modules/relations/public-api.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './relations.component';
-export * from './relations.module';
