+++ projects/natural/src/lib/modules/search/public-api.ts
@@ -8,7 +8,6 @@ export {FlagFacet, DropdownFacet, Facet, NaturalSearchFacets} from './types/face
 export {NaturalSearchSelections, NaturalSearchSelection} from './types/values';
 export {NaturalDropdownRef} from './dropdown-container/dropdown-ref';
 export {NATURAL_DROPDOWN_DATA} from './dropdown-container/dropdown.service';
-export {NaturalSearchModule} from './search.module';
 export {toGraphQLDoctrineFilter} from './classes/graphql-doctrine';
 export {fromUrl, toUrl} from './classes/url';
 export * from './classes/transformers';
