--- projects/natural/src/lib/modules/search/search.module.spec.ts
@@ -1,7 +1,0 @@
-import {NaturalSearchModule} from './search.module';
-
-describe('NaturalSearchModule', () => {
-    it('should work', () => {
-        expect(new NaturalSearchModule()).toBeDefined();
-    });
-});
