+++ projects/natural/src/lib/modules/select/public-api.ts
@@ -5,4 +5,3 @@
 export * from './select-enum/select-enum.component';
 export * from './select-hierarchic/select-hierarchic.component';
 export * from './select/select.component';
-export * from './select.module';
