+++ projects/natural/src/lib/modules/stamp/public-api.ts
@@ -2,5 +2,4 @@
  * Public API Surface of natural
  */
 
-export * from './stamp-module.module';
 export * from './stamp.component';
