+++ projects/natural/src/lib/modules/table-button/public-api.ts
@@ -3,4 +3,3 @@
  */
 
 export * from './table-button.component';
-export * from './table-button.module';
