--- projects/natural/src/lib/modules/table-button/table-button.module.ts
@@ -1,13 +1,0 @@
-import {CommonModule} from '@angular/common';
-import {NgModule} from '@angular/core';
-import {MatButtonModule} from '@angular/material/button';
-import {RouterModule} from '@angular/router';
-import {NaturalIconModule} from '../icon/icon.module';
-import {NaturalTableButtonComponent} from './table-button.component';
-
-@NgModule({
-    declarations: [NaturalTableButtonComponent],
-    imports: [CommonModule, RouterModule, MatButtonModule, NaturalIconModule],
-    exports: [NaturalTableButtonComponent],
-})
-export class NaturalTableButtonModule {}
