+++ src/app/file/file.component.spec.ts
@@ -2,7 +2,7 @@ import {waitForAsync, ComponentFixture, TestBed} from '@angular/core/testing';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {RouterTestingModule} from '@angular/router/testing';
 import {
-    NaturalAlertModule,
+    NaturalModule,
     NaturalColumnsPickerModule,
     NaturalCommonModule,
     NaturalDetailHeaderModule,
@@ -36,7 +36,7 @@ describe('Demo FileComponent', () => {
                     RouterTestingModule,
                     BrowserAnimationsModule,
                     MaterialModule,
-                    NaturalAlertModule,
+                    NaturalModule,
                     NaturalStampModule,
                     NaturalSelectModule,
                     NaturalSidenavModule,
