+++ src/app/home/home.component.spec.ts
@@ -2,7 +2,7 @@ import {waitForAsync, ComponentFixture, TestBed} from '@angular/core/testing';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {RouterTestingModule} from '@angular/router/testing';
 import {
-    NaturalAlertModule,
+    NaturalModule,
     NaturalColumnsPickerModule,
     NaturalCommonModule,
     NaturalDetailHeaderModule,
@@ -33,7 +33,7 @@ describe('Demo HomeComponent', () => {
                     RouterTestingModule,
                     BrowserAnimationsModule,
                     MaterialModule,
-                    NaturalAlertModule,
+                    NaturalModule,
                     NaturalStampModule,
                     NaturalSelectModule,
                     NaturalSidenavModule,
